+++ runner/executor.go
@@ -1,4 +1,4 @@
-package internal
+package runner
 
 import "fmt"
